E-commer-project-main/requirements.txt
Django==3.2.9
django-filter==21.1
djangorestframework==3.12.4
pytz==2021.3
PyJWT==v1.7.1
redis==4.3.6
djangorestframework-simplejwt==4.4.0
psycopg2-binary>=2.8,<3.0
celery>=5.0,<6.0